+++ db/migrate/20131016062210_charity_grants.rb
@@ -1,17 +1,17 @@
 class CharityGrants < ActiveRecord::Migration
   def change
     create_table :charity_grants do |t|
-      t.integer :charity_id
-      t.integer :charity_group_id
-      t.integer :donor_id
-      t.decimal   :shares_subtracted
-      t.integer :transaction_id
-      t.date    :date
-      t.decimal   :transaction_fee
-      t.decimal   :giv2giv_fee
-      t.decimal   :gross_amount
-      t.decimal   :grant_amount
-      t.string  :status
+      t.integer   :charity_id
+      t.integer   :charity_group_id
+      t.integer   :donor_id
+      t.float   :shares_subtracted
+      t.integer   :transaction_id
+      t.date      :date
+      t.float   :transaction_fee
+      t.float   :giv2giv_fee
+      t.float   :gross_amount
+      t.float   :grant_amount
+      t.string    :status
       t.timestamps
     end
   end
